requirements.txt
certifi>=2018.4.16
chardet>=3.0.4
facebook-sdk>=2.0.0
idna>=2.7
oauthlib>=2.1.0
PySocks>=1.6.8
requests>=2.19.1
requests-oauthlib>=1.0.0
six>=1.11.0
tweepy>=3.6.0
urllib3>=1.23
